+++ postnormalism/schema/function.py
@@ -1,11 +1,14 @@
 from dataclasses import dataclass
 
-from .schema_item import SchemaItem
+from .database_item import DatabaseItem
 
 
 @dataclass(frozen=True)
-class Function(SchemaItem):
+class Function(DatabaseItem):
     """
     A data class for functions.
     """
-    pass
+
+    def __post_init__(self):
+        function_name_pattern = r'CREATE\s+(?:OR\s+REPLACE\s+)?(?:TEMP\s+)?(?:FUNCTION)\s+(?:IF\s+NOT\s+EXISTS\s+)?(?:\w+\.)?(\w+)'
+        super().__post_init__(function_name_pattern)